requirements.txt
pandas
numpy
streamlit
matplotlib
python-dotenv
google-generativeai